final_vqa_project-2 2/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
datasets>=2.18.0
pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.4.0
tqdm>=4.66.0
streamlit>=1.35.0
accelerate>=0.28.0
sentencepiece>=0.2.0
